refactor and move common code to RecordHandler
